refactor: reorganize feedback similarity types, constants, processor, and update imports for consistency
